This is just for testing
While working on the getcondition function I needed a way to test stuff out. This commit is just so that I can continue working on this test function while I'm travelling. Can therefore be ignored, unless you want to understand it further, in which case you are welcome to read on