ICImplementation + Intro / Next Button Change
Added Page where the information consent is displayed;
Added A checkbox for the information consent;
Added Zustand for accepting the conditions of information consent;

Time Zustand got a new variable for tracking when the button on Information consent was pressed;

Changed the page-switch in the index.ts and added the information consent page there.
Also made changes to the intro page where i removed unnecessary Parameters.

Reworked NextPageButton added 2 paremters for reusability:
 - Added timeVar -> Is used for setting the right time;
 - Added buttonText -> Variable for changing the button Text