Removed delay on the PreStudyPage
The next page method has been removed from the try scope. This should make the delay disappear.
I'm not sure if this is the best way, as we now have no feedback as to whether the condition assignment really worked.
To be on the safe side, we should check in the task whether this has worked.